feat: Login page folder restructure and refactor
