+++ src/pages/Login/CardBody/LoginCardBody.test.jsx
@@ -1,7 +1,7 @@
 import { matchMediaConfig } from '@/utils/testUtils'
 import { fireEvent, render, screen } from '@testing-library/react'
 import { describe, test } from 'vitest'
-import LoginCardBody from './LoginCardBody'
+import LoginCardBody from '.'
 
 matchMediaConfig()
