+++ src/pages/Login/CardFooter/index.jsx
@@ -1,8 +1,8 @@
-const LoginCardFooter = () => (
+const CardFooter = () => (
 	<p className='text-center w-2/3 mx-auto'>
 		This product is licensed for Dataon Corporation <br />© 1999 - 2022 Dataon
 		Technologies, All Rights Reserved
 	</p>
 )
 
-export default LoginCardFooter
+export default CardFooter
