+++ src/pages/Login/Wrapper/LoginWrapper.test.jsx
@@ -1,7 +1,7 @@
 import bgImage from '@/assets/background.svg'
 import { render, screen } from '@testing-library/react'
 import { describe, expect, test } from 'vitest'
-import LoginWrapper from './LoginWrapper'
+import LoginWrapper from '.'
 
 describe('Login Card header test', () => {
 	test('should have the correct background image', () => {
